restructure `moss-collection` crate
Move "api" and "worktree" folders to the top level
