+++ crates/moss-collection/src/api/stream_entries_by_prefixes.rs
