--- crates/moss-collection/src/collection/utils.rs
@@ -1,1 +1,0 @@
-
