--- crates/moss-collection/src/constants.rs
@@ -1,13 +1,0 @@
-pub(crate) const REQUESTS_DIR: &'static str = "requests";
-
-pub(crate) const ANY_ENTRY_SPEC_FILE_EXT: &'static str = "sapic";
-pub(crate) const FOLDER_ENTRY_DIR_EXT: &'static str = "request";
-
-pub(crate) const GET_ENTRY_SPEC_FILE: &'static str = "get.sapic";
-pub(crate) const POST_ENTRY_SPEC_FILE: &'static str = "post.sapic";
-pub(crate) const PUT_ENTRY_SPEC_FILE: &'static str = "put.sapic";
-pub(crate) const DELETE_ENTRY_SPEC_FILE: &'static str = "delete.sapic";
-pub(crate) const GRAPHQL_ENTRY_SPEC_FILE: &'static str = "gql.sapic";
-pub(crate) const GRPC_ENTRY_SPEC_FILE: &'static str = "grpc.sapic";
-
-pub(crate) const FOLDER_ENTRY_SPEC_FILE: &'static str = "folder.sapic";
