+++ crates/moss-collection/src/lib.rs
@@ -1,4 +1,5 @@
+pub mod api;
 pub mod collection;
-pub mod constants;
 pub mod kdl;
 pub mod models;
+pub mod worktree;
