+++ crates/moss-collection/src/worktree.rs
