+++ crates/moss-collection/src/worktree/snapshot.rs
